ASySD: Automated Citation Deduplication

The Automated Systematic Search Deduplicator ('ASySD') includes functions to automatically detect and remove duplicate citations from literature searches. For details, see our paper "The Automated Systematic Search Deduplicator (ASySD): a rapid, open-source, interoperable tool to remove duplicate citations in biomedical systematic reviews. Hair K / Bahor Z / Macleod M / Liao J / Sena E (2023) doi:10.1186/s12915-023-01686-z .

See also

Author

Maintainer: Kaitlyn Hair kaitlyn.hair@ed.ac.uk (ORCID) [copyright holder]

Authors:

Lukas Wallrich lukas.wallrich@gmail.com (ORCID)